chore: move rules to a base package
