+++ .changeset/eslint-config-private-base-split.md
@@ -1,0 +1,5 @@
+---
+'@seedcord/eslint-config': patch
+---
+
+Split the config content into a private bundled base package. No consumer-visible change.
