+++ .changeset/pre.json
@@ -2,24 +2,25 @@
   "mode": "pre",
   "tag": "next",
   "initialVersions": {
-    "@seedcord/docs": "0.1.5",
-    "@seedcord/home": "0.0.1",
-    "@seedcord/mock": "0.2.22",
     "@seedcord/cli": "0.3.1",
+    "@seedcord/core": "0.0.1",
+    "@seedcord/docs": "0.1.5",
     "@seedcord/docs-engine": "0.3.1",
     "@seedcord/docs-generator": "0.2.4",
     "@seedcord/errors": "0.2.1",
     "@seedcord/eslint-config": "1.4.3",
+    "@seedcord/eslint-config-base": "0.0.1",
     "@seedcord/eslint-plugin": "0.0.1",
+    "@seedcord/home": "0.0.1",
+    "@seedcord/mock": "0.2.22",
     "@seedcord/plugins": "0.7.2",
-    "seedcord": "0.15.0",
     "@seedcord/services": "0.8.2",
     "@seedcord/tsconfig": "2.0.1",
     "@seedcord/tsdown-config": "2.0.1",
     "@seedcord/types": "0.7.1",
     "@seedcord/ui": "0.2.0",
     "@seedcord/utils": "0.7.0",
-    "@seedcord/core": "0.0.1"
+    "seedcord": "0.15.0"
   },
   "changesets": [
     "bus-hmr-registration",
