+++ .github/labels.yml
@@ -24,6 +24,9 @@
   color: 629bbc
 - name: '📦 eslint-config'
   color: 6278bc
+- name: '📦 eslint-config-base'
+  color: 6278bc
+  description: Internal ESLint config content bundled into eslint-config
 - name: '📦 eslint-plugin'
   color: 6272bc
   description: ESLint plugin for seedcord and discord.js bots
