+++ apps/docs/package.json
@@ -49,9 +49,9 @@
         "@types/react-dom": "catalog:react",
         "@vitest/coverage-v8": "catalog:test",
         "eslint": "^9.39.4",
-        "eslint-config-next": "catalog:deps",
-        "eslint-plugin-jsx-a11y": "catalog:deps",
-        "eslint-plugin-react-compiler": "catalog:deps",
+        "eslint-config-next": "catalog:eslint",
+        "eslint-plugin-jsx-a11y": "catalog:eslint",
+        "eslint-plugin-react-compiler": "catalog:eslint",
         "jsdom": "catalog:test",
         "prettier-plugin-tailwindcss": "catalog:deps",
         "typescript": "catalog:peer",
