+++ apps/home/package.json
@@ -29,9 +29,9 @@
         "@types/react": "catalog:react",
         "@types/react-dom": "catalog:react",
         "eslint": "^9.39.4",
-        "eslint-config-next": "catalog:deps",
-        "eslint-plugin-jsx-a11y": "catalog:deps",
-        "eslint-plugin-react-compiler": "catalog:deps",
+        "eslint-config-next": "catalog:eslint",
+        "eslint-plugin-jsx-a11y": "catalog:eslint",
+        "eslint-plugin-react-compiler": "catalog:eslint",
         "prettier": "catalog:deps",
         "prettier-plugin-tailwindcss": "catalog:deps",
         "tailwindcss": "catalog:deps",
