+++ packages/eslint-config-base/src/constants.ts
