+++ packages/eslint-config-base/src/mdx.ts
@@ -2,15 +2,7 @@ import * as mdx from 'eslint-plugin-mdx';
 
 import type { Linter } from 'eslint';
 
-/**
- * Builds the flat-config block that turns on MDX support.
- *
- * Deliberately omits the `mdx/remark` rule and its remark processor: that bridge lints
- * markdown *prose* via remark-lint, which overlaps with markdownlint and needs a separate
- * remark config to do anything. Prose linting stays out of scope here.
- *
- * @param files - File globs the block applies to (typically every `.mdx` file).
- */
+// omits mdx/remark on purpose, that bridge lints markdown prose and needs its own remark config
 export function mdxBlock(files: string[]): Linter.Config {
     const { languageOptions, plugins } = mdx.flat;
