+++ packages/eslint-config-base/src/options.ts
