+++ packages/eslint-config-base/src/rules/general-rules.ts
