+++ packages/eslint-config-base/src/rules/index.ts
