+++ packages/eslint-config-base/src/rules/tsdoc-rules.ts
