+++ packages/eslint-config-base/src/rules/typescript-rules.ts
