+++ packages/eslint-config-base/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "extends": "@seedcord/tsconfig/node",
+    "compilerOptions": {
+        "outDir": "./dist",
+        "rootDir": ".",
+        "allowJs": true,
+        "checkJs": false
+    },
+    "include": ["./src", "eslint.config.mjs", "./tsdown.config.ts"],
+    "exclude": ["node_modules", "dist"]
+}
