+++ packages/eslint-plugin/eslint.config.mjs
@@ -1,4 +1,4 @@
-import createConfig from '@seedcord/eslint-config';
+import createConfig from '@seedcord/eslint-config-base';
 
 export default [
     ...createConfig({ tsconfigRootDir: import.meta.dirname }),
