+++ pnpm-workspace.yaml
@@ -31,10 +31,7 @@ catalogs:
     chalk: ^5.6.2
     discord-api-types: ^0.38.49
     envapt: ^6.0.2
-    eslint: ^10.6.0
-    eslint-config-next: ^16.2.9
-    eslint-plugin-jsx-a11y: ^6.10.2
-    eslint-plugin-react-compiler: 19.1.0-rc.2
+    lodash.merge: ^4.6.2
     lucide-react: ^1.21.0
     motion: ^12.40.0
     next: ^16.2.9
@@ -46,6 +43,24 @@ catalogs:
     tsx: ^4.22.4
     type-fest: ^5.7.0
     vite: ^8.0.16
+  eslint:
+    '@types/eslint-plugin-security': ^3.0.1
+    '@typescript-eslint/parser': ^8.61.1
+    eslint: ^10.6.0
+    eslint-config-next: ^16.2.9
+    eslint-config-prettier: ^10.1.8
+    eslint-import-resolver-typescript: ^4.4.5
+    eslint-plugin-better-tailwindcss: ^4.6.0
+    eslint-plugin-import-x: ^4.17.1
+    eslint-plugin-jsx-a11y: ^6.10.2
+    eslint-plugin-mdx: ^3.8.1
+    eslint-plugin-prettier: ^5.5.6
+    eslint-plugin-react-compiler: 19.1.0-rc.2
+    eslint-plugin-security: ^4.0.1
+    eslint-plugin-tailwind-canonical-classes: ^1.3.3
+    eslint-plugin-tsdoc: ^0.5.2
+    eslint-plugin-unicorn: ^70.0.0
+    typescript-eslint: ^8.61.1
   peer:
     discord.js: ^14.26.4
     mongoose: ^9.7.1
